优化wire注入
