+++ app/internal/pkg/database/database.go
@@ -2,22 +2,17 @@ package database
 
 import (
 	"github.com/google/wire"
-	"github.com/zeromicro/go-zero/core/stores/sqlx"
 )
 
 var Provider = wire.NewSet(
+	NewDatabase,
 	NewDefault,
 )
 
-// Config 数据库连接信息
-type Config struct {
-	Default string // 默认数据库
+type Database struct {
+	Default *Default
 }
 
-type Default struct {
-	sqlx.SqlConn
-}
-
-func NewDefault(c *Config) *Default {
-	return &Default{sqlx.NewMysql(c.Default)}
+func NewDatabase(Default *Default) *Database {
+	return &Database{Default: Default}
 }
