+++ app/internal/pkg/database/types.go
@@ -1,0 +1,6 @@
+package database
+
+// Config 数据库连接信息
+type Config struct {
+	Default string // 默认数据库
+}
