+++ app/internal/pkg/redis/redis.go
@@ -2,27 +2,17 @@ package redis
 
 import (
 	"github.com/google/wire"
-	"github.com/zeromicro/go-zero/core/stores/redis"
 )
 
 var Provider = wire.NewSet(
+	NewRedis,
 	NewDefault,
 )
 
-type Config struct {
-	Host string
-	Type string
-	Pass string
+type Redis struct {
+	Default *Default
 }
 
-type Default struct {
-	*redis.Redis
-}
-
-func NewDefault(c *Config) *Default {
-	return &Default{redis.MustNewRedis(redis.RedisConf{
-		Host: c.Host,
-		Type: c.Type,
-		Pass: c.Pass,
-	})}
+func NewRedis(Default *Default) *Redis {
+	return &Redis{Default: Default}
 }
