+++ app/internal/pkg/redis/types.go
@@ -1,0 +1,7 @@
+package redis
+
+type Config struct {
+	Host string
+	Type string
+	Pass string
+}
