+++ app/internal/svc/service_context.go
@@ -5,6 +5,7 @@ import (
 	"go-zero-box/app/internal/config"
 	"go-zero-box/app/internal/middleware"
 	"go-zero-box/app/internal/model"
+	"go-zero-box/app/internal/pkg/database"
 	"go-zero-box/app/internal/pkg/redis"
 	"go-zero-box/app/internal/services"
 )
@@ -17,13 +18,14 @@ var Provider = wire.NewSet(
 )
 
 type ServiceContext struct {
-	RedisDefault *redis.Default
-	Config       *config.Config
-	Model        *model.Model
-	Service      *services.Services
+	Redis    *redis.Redis
+	Database *database.Database
+	Config   *config.Config
+	Model    *model.Model
+	Service  *services.Services
 	*middleware.Middleware
 }
 
-func NewServiceContext(redisDefault *redis.Default, config *config.Config, model *model.Model, service *services.Services, middleware *middleware.Middleware) *ServiceContext {
-	return &ServiceContext{RedisDefault: redisDefault, Config: config, Model: model, Service: service, Middleware: middleware}
+func NewServiceContext(redis *redis.Redis, database *database.Database, config *config.Config, model *model.Model, service *services.Services, middleware *middleware.Middleware) *ServiceContext {
+	return &ServiceContext{Redis: redis, Database: database, Config: config, Model: model, Service: service, Middleware: middleware}
 }
